% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble plot of CMap output table}
\usage{
bubble_plot(path, output_path=NULL)
}
\arguments{
\item{path}{path of the excel file}

\item{output_path}{path for the experiment output folder (default=NULL)}
}
\value{
ggplot object - bubble plot
}
\description{
This function allows the user to represent the CMap results (broadinstitute)
under the form of a bubble plot representing statistics and cell lines.
}
\examples{
# bubble_plot(path, output_path=NULL)
}
